# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/carts:
    get:
      tags:
        - Carts
      summary: List Carts
      description: Get a paginated list of Carts with optional filtering and sorting
      operationId: listCarts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Carts
      summary: Create Cart
      description: Create a new Cart
      operationId: createCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCartRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/cart_lines:
    get:
      tags:
        - CartLines
      summary: List CartLines
      description: Get a paginated list of CartLines with optional filtering and sorting
      operationId: listCartLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CartLines
      summary: Create CartLine
      description: Create a new CartLine
      operationId: createCartLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCartLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Cart:
      type: object
      required:
        - id
        - website_id
        - visitor_id
        - state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        visitor_id:
          type: string
          format: uuid
        portal_user_id:
          nullable: true
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        state:
          type: string
        coupon_code:
          nullable: true
          type: string
          maxLength: 60
        delivery_carrier_id:
          nullable: true
          type: string
          format: uuid
        placed_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateCartRequest:
      type: object
      required:
        - website_id
        - visitor_id
        - state
        - metadata
      properties:
        website_id:
          type: string
          format: uuid
        visitor_id:
          type: string
          format: uuid
        portal_user_id:
          nullable: true
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        state:
          type: string
        coupon_code:
          nullable: true
          type: string
          maxLength: 60
        delivery_carrier_id:
          nullable: true
          type: string
          format: uuid
        placed_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CartList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Cart'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CartLine:
      type: object
      required:
        - id
        - cart_id
        - item_id
        - quantity
        - metadata
      properties:
        id:
          type: string
          format: uuid
        cart_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        metadata:
          type: object
...
    CreateCartLineRequest:
      type: object
      required:
        - cart_id
        - item_id
        - quantity
        - metadata
      properties:
        cart_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        metadata:
          type: object
...
    CartLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []